requirements.txt
fastapi
strawberry-graphql[fastapi]
sqlmodel
sqlalchemy
pydantic
uvicorn
pytest 